Car Rental App
I developed the project using Firebase, and if you want to run the project, you can do so by using a Firebase Firestore database. Please note that the column names must be the same 😊😊. Additionally, the warning messages and column names are written in my native language, Turkish. If there are parts you don't understand, you can check the comment lines or use translation tools to overcome this issue.